chore: update lodestar script to properly forward exit signals (#7383)
**Motivation**

Users might use the lodestar script inside docker or with a process
manager like systemd (https://github.com/ChainSafe/lodestar/issues/7378)
and in those cases we need to make sure Lodestar runs as primary process
and not the shell script as otherwise exit signals are not properly
forwarded or in case of systemd it does not wait for Lodestar to shut
down.

**Description**

Update lodestar script to properly forward exit signals by execution the
main node process via `exec`. This shouldn't change anything in regard
to how we use the script during development.