chore: fix prune history panels (#7442)
Noticed the queries are not working as the `_bucket` suffix is missing.
Also did some cosmetic changes and moved the panels a bit further down
in the sync dashboard.


![image](https://github.com/user-attachments/assets/37489fd6-1062-46c2-9d9e-31f4c460a9db)


We might also wanna reconsider buckets as fetching keys seems to take >
1 second in a few cases. I mentioned a [possible
solution](https://discord.com/channels/593655374469660673/1337188931489239072/1338499158566375464)
to improve the fetch time.