feat: rename pre/post fork designations and fork guards (#7441)
**Motivation**

Start of resolution to #7183. Driven by peerDAS branch. `isForkBlobs` is
not longer semantically correct in `fulu` because that branch does not
have blobs, it has columns. Is the first major feature that was removed
so the types/typeguard naming semantics we were using kinda broke.
Updated to reflect pre/post fork instead of post"Feature".

**Description**

Rename pre/post fork names and type guards.

---------

Co-authored-by: Nico Flaig <nflaig@protonmail.com>