chore: remove outdated reasoning of the API definitions comment (#7475)
**Motivation**

Forgot to update this as part of
https://github.com/ChainSafe/lodestar/pull/6749, but adding new
endpoints is much simpler / cleaner now and everything is inferred from
`Endpoint` type definition when adding a new route, it's not possible to
forget updating something as typescript will complain (and also spec
tests). So instead of updating this I figured it's best to just remove
it, looking at the code is much easier to figure out what do to than
having to read some text I would put there, eg. see example
https://github.com/ChainSafe/lodestar/pull/7472.

**Description**

Remove outdated reasoning of the API definitions comment