fix: reject a consolidation request coming from a validator with BLS withdrawal credentials (#7568)
**Motivation**

In order to comply with the consensus-specs of Electra, when processing
a consolidation request we should check that the source validator does
not have withdrawal credentials with a BLS prefix.

**Description**

The prefix of the withdrawal credentials of the source validator that is
targeted by the consolidation request should either be
`ETH1_ADDRESS_WITHDRAWAL_PREFIX` or `COMPOUNDING_WITHDRAWAL_PREFIX`.
The relevant function is
[`process_consolidation_request`](https://github.com/ethereum/consensus-specs/blob/dev/specs/electra/beacon-chain.md#new-process_consolidation_request)
and as expected verifies the prefix of the withdrawal credentials of the
source validator:
```python
    # Verify source withdrawal credentials
    has_correct_credential = has_execution_withdrawal_credential(source_validator)
    is_correct_source_address = (
        source_validator.withdrawal_credentials[12:] == consolidation_request.source_address
    )
    if not (has_correct_credential and is_correct_source_address):
        return
```
Currently, a validator could create a deposit transaction with BLS
withdrawal credentials ending with an eth1 that they control in order to
make the consolidation request from this address later.
Since Lodestar will process this request and will add the request, it
will have a different state root from other clients, effectively
performing a chain split as showcased by the failing test at the first
commit.

---------

Co-authored-by: Nico Flaig <nflaig@protonmail.com>