feat: make committee_index optional for post-electra produce attestation data calls (#7626)
**Motivation**

We already ignore `committee_index` in post-electra produce attestation
data calls and always set it to 0 in returned `AttestationData`, however
we can be less strict on the beacon node side and only reject api calls
that do not provide the committee index pre-electra.

**Description**

This changes the way we enforce that `committee_index` is provided by
caller, it moves the check from static schema validation to fork-based
validation which allows us to handle requests without the committee
index for post-electra.

The validator client will still pass the committee index in any case as
per spec this is a required property as of right now. However, there was
some discussion around this in
https://github.com/ethereum/beacon-APIs/pull/511 and a long-term plan
could be to make the `committee_index` optional post-electra on the spec
as well but first beacon nodes need to be more relaxed on how they
enforce the parameter requirement.