chore: reduce time to retain registered validators in monitor to 1 hour (#7668)
**Motivation**

I don't see a good reason why we need to retain validators we monitor
for longer than an hour. The registration is updated every epoch (~6.4
minutes) if the validator client is connected and downtime due to
maintenance should only be a few minutes and at most 2-3 epochs if
doppelganger protection is enabled (and not even then since we have
https://github.com/ChainSafe/lodestar/pull/6012) or keys are moved
around.

This long retention period was noticable during holesky rescue as we
offered our beacon node to the community but due to the fact that the
validator monitor keeps track of validators for 12 hours even if those
are not even connected anymore it made it harder to estimate how many
validators are actually connected and managed by the beacon node(s).

There was also a [complaint on
discord](https://discord.com/channels/593655374469660673/593655641445367808/1248614428329513072)
regarding this before.

**Description**

Reduce time to retain registered validators in monitor from previously
12 hours to only 1 hour.

We can re-evaluate this number in the future if it turns out to be
problematic but I don't see a scenario right now where it would be
useful to retain validators for longer which are not even connected
anymore.