docs: add slashing protection information (#7689)
**Motivation**

It is not perfectly clear how slashing protection works in lodestar.
Therefore a short section should be added to explain some details.

**Description**

This PR adds a short section to the page `Starting a Validator Client`
which describes slashing protection behavior and how to migrate to and
from a different client.

---------

Co-authored-by: towo <Tobias.Wohland@telekom.de>
Co-authored-by: Nico Flaig <nflaig@protonmail.com>