feat: peerDAS - implement validator custody (#7640)
**Motivation**

@hughy and I have a basic implementation of [validator
custody](https://github.com/ethereum/consensus-specs/blob/dev/specs/fulu/validator.md)
for peerDAS.

We're planning to do more testing on this, but would appreciate a review
on the architecture since we're still pretty new!

This relates to #7632 - If it merges, we'll update our PR to account for
it.

**Description**

* Centralizes custody values like `sampledGroups` and `custodyGroups`
into `CustodyConfig`. `CustodyConfig` is now treated as a singleton.
* Creates a new `advertisedGroupCount` in `CustodyConfig`, used for the
custody group count in the node's metadata/ENR.
* Adds `setSamplingGroupCount` and `setAdvertisedGroupCount` to
NetworkCore API. Updated by an `EventEmitter` on `CustodyConfig`.
* Adds LocalValidatorRegistry to track connected validators.
* Updates custody requirement in `chain.onForkChoiceFinalized`.

**Not Included**

I'll open separate issues for these if we're okay merging this PR
without them.

* Backfilling groups when the target custody group count increases
* Handling changes in other peers' custody group counts
* Race conditions around group count changing during syncing

<!-- A clear and concise general description of the changes of this PR
commits -->

<!-- If applicable, add screenshots to help explain your solution -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #7619

---------

Co-authored-by: Hugh Cunningham <hugh.e.cunningham@gmail.com>