fix: peerDAS - support MetadataV3 before fulu (#7700)
**Motivation**

Running a cluster of three nodes in Kurtosis, was seeing this error
before the Fulu fork:

```
[cl-1-lodestar-geth] Error: Request to send to protocol /eth2/beacon_chain/req/metadata/3/ssz_snappy but it has not been declared
[cl-1-lodestar-geth]     at ReqRespBeaconNode.sendRequest (file:///usr/app/packages/reqresp/lib/ReqResp.js:104:23)
```

The error is because the client is attempting to send Metadatav3
requests. There's a check in the sender that prevents sending messages
to versions that the client hasn't registered, and Metadatav3 is not
registered until the fork.

[The Fulu
spec](https://github.com/ethereum/consensus-specs/blob/dev/specs/fulu/p2p-interface.md#getmetadata-v3)
says to follow the same semantics as Altair, which allows registering
the new metadata endpoint before the fork.

I think that's a good idea anyway, to allow clients to prioritize peers
pre-fork based on their expected custody groups.

**Description**

* Updates MetadataV3 to always be registered
* Updates MetadataV2 to unregister at Fulu fork, like MetadataV1 does at
Altair