fix: always add CGC to ENR (#7738)
**Motivation**

The MetadataController currently sets CGC on the ENR in two cases:

* When the node is first started after the Fulu epoch
* Any time `setAdvertisedGroupCount` is called

We should at least ensure CGC is set when crossing the Fulu fork
boundary so that nodes are aware of our CGC (for example, if we're set
as a supernode, nodes would currently assume we only custody 4 columns).

Additionally, I think we should have CGC available prior to the fork
boundary in case nodes want to use it to prioritize peers before the
fork. We already do this for Metadata by making the MetadataV3 endpoint
available prior to the fork.

<!-- Why is this PR exists? What are the goals of the pull request? -->

**Description**

* Removes the condition around CGC in the upstreamValues function so
that CGC is sent to ENR regardless of the epoch at node start.
* Adds a test to make sure CGC is always set by upstreamValues