fix: do not trigger builder circuit breaker if clock slot < slots present (#7481)
**Motivation**

We currently disable the builder flow for the most part of epoch 0
because the circuit breaker is triggered and we fall back to local block
building. However the current check does not make much sense if there
were not enough slots yet where proposal could be included on-chain.

**Description**

Do not trigger builder circuit breaker if clock slot < slots present
(based on fork choice)