feat: make aggregate with randomness async again (#7761)
**Motivation**

Gnosis analysis shows that there is a lot of time spent on main thread
for running aggregateWithRandomness on main thread. Now that libp2p@v2
is more stable and so are some of the other conditions that triggered
our revert last time, looking to check this again to see what the
metrics look like for running not on main thread.

---------

Co-authored-by: Nico Flaig <nflaig@protonmail.com>