chore: fix broken panels due to missing historical state worker metrics (#7826)
**Motivation**

Since https://github.com/ChainSafe/lodestar/pull/7799, the historical
state worker is only enabled if `--serveHistoricalState` flag is set
which means that metrics might not exist. This is not a problem for
dedicated panels / dashboard for the historical state worker / regen but
it shouldn't break other metric panels that include the metrics as part
of a sum.

**Description**


Fix broken panels by adding `or vector(0)` to
`lodestar_historical_state_worker_*` metrics to handle the case if those
metrics are missing due to historical state work being disabled