chore: restore dashboard variable names and inputs during download (#7872)
**Motivation**

Closes https://github.com/ChainSafe/lodestar/issues/7840, it seems that
newer versions of grafana replace variables with values picked during
manual import. This means if an external contributor that edits a
dashboard and follows our steps in contribution guideline it will still
keep the hard coded values and requires to manually revert them, which
is not a big deal but we should still make sure that external
contributor have as close to seamless experience when contributing to
dashboard as we do using the cloud instance with provisioned dashboards.

**Description**

Restore dashboard variable names and inputs during download