fix: fix proposer boost reorg by suppressing fcu call during import (#7810)
We need to suppress fcu call during gossip block import when the
incoming block is weak. This is because once notified EL, it will treat
the new block as head and will not reorg back to parent if we try to
reorg.

- Add `shouldOverrideForkChoiceUpdate`
- Call `shouldOverrideForkChoiceUpdate` during gossip block import, and
skip fcu call if result is true
- Modify `predictProposerHead` to be a simple wrapper around
`shouldOverrideForkChoiceUpdate`


Closes #7235

cc. @twoeths @nflaig

---------

Co-authored-by: Nico Flaig <nflaig@protonmail.com>