feat: introduce finalProcessEpoch() (#8086)
**Motivation**

~~as found in #8082, there is cyclic dependency between
`afterProcessEpoch` and `upgradeState*` like `upgradeStateToElectra`~~

this PR improves a couple of things:

- to prepare for the future: future contributor can just decide where to
put logic at and it's straightforward to them. Cyclic dependency does
not happen here because the logic of computing proposers is unchanged
for fulu, but it may happen for other fields in the future
- at fork boundary, (epoch 0 of fulu), we implicitly run pre-fork logic,
only run post-fork logic from epoch 1 and it's not great. In epoch
transition we should make everything explicit to make it easier to
maintain
- performance: we compute proposers for current epoch and next epoch
twice at epoch 0 of fulu

**Description**

- implement finalProcessEpoch(), this is to compute epoch data that
depends on `upgradeState*`
- call this after `upgradeState*()`

---------

Co-authored-by: Tuyen Nguyen <twoeths@users.noreply.github.com>
Co-authored-by: Nico Flaig <nflaig@protonmail.com>