fix: the data column sidecars behavior for unavailable sidecars (#8251)
**Motivation**

The check was missed on the length of the objects. 

**Description**

- The check was required to have on the length of the array
- Earlier we were returning `T[] | null`, later changed to `(T |
null)[]`, and some was missed the check and also weird that linter
didn't found it.

**Steps to test or reproduce**

- Run all tests


**Notes**

As stated by @nflaig the specs says as below:

```
in the deneb spec it's pretty clear that we should skip in range requests
> Slots that do not contain known blobs MUST be skipped, mimicking the behaviour of the BlocksByRange request. Only response chunks with known blobs should therefore be sent.

similar note for columns
> Slots that do not contain known data columns MUST be skipped, mimicking the
> behaviour of the `BlocksByRange` request. Only response chunks with known data
> columns should therefore be sent.

```

So if we throw error during the serving of data columns, we might miss
serving the one's which we have.