docs: clarify ipv6 details in networking page (#8368)
**Motivation**

After resolving issues with the ipv6 dual stack port bug in v1.34.1, it
was noted that we should make it clear our default p2p port for both
IPv4 and IPv6 is now the same at port `9000`.

**Description**

When we introduced IPv6 support in #5758 , we used to default the port
of `9090`. When supporting dual stack option on the same port via #8190
, we should make it clear in documentation this is now by default also
on port `9000`.