feat: include commit hash in network lodestar version (#8605)
We are already including the version (eg. `lodestar/v1.36.0`), there
doesn't seem to be much of a benefit in terms of security to not include
the commit hash as well and it helps debugging especially in early
testnets or release candidates as there the version number is the same
while it might run a different commit.

We do have the `--private` flag to avoid including any information about
the client on p2p.